fix: 适配实际数据集格式，修复数据管道 7 个 bug
实际数据集与数据字典存在多处差异（rating 为文本而非数值、metadata
仅有 tags 无 genres 等），修复管道中因假设不匹配导致的运行时错误：

- cleaner: 检测 rating dtype，文本型跳过数值中位数填充和异常值检测
- builder: rating 文本时回退使用 positive_ratio/100 作为数值特征
- builder: app_id/user_id 不在索引时自动 set_index，保证 merge 正确
- builder: 去除特征矩阵中的 date 列，防止 sklearn DTypePromotionError
- builder: genres 全部为空时 num_genres 回退为 num_tags
- eda_plots: 评分分布支持文本型（柱状图）和数值型（直方图）
- eda_plots: genres 为空时回退使用 tags 绘制 Top 20 类别图
- model_plots: PDP 前将 X 转为 float，避免 int64 列报错
- main: extreme 结果键名从平铺改为嵌套访问，补充 np import

88 tests 全部通过，管道 8 阶段完整运行成功。

+++ src/visualization/model_plots.py
@@ -123,7 +123,7 @@ def plot_partial_dependence(model, X, features, save_path: str) -> None:
     save_path : str
     """
     display = PartialDependenceDisplay.from_estimator(
-        model, X, features, grid_resolution=20
+        model, X.astype(float), features, grid_resolution=20
     )
     display.figure_.tight_layout()
     display.figure_.savefig(save_path, dpi=150)
